refactor: move workflow mechanics beside steps
Keep RepresentativeWorkflowRunner focused on orchestration by moving archive, restore, conflict, and archive-tier mechanics into step-adjacent helpers under Workflows/Steps. Preserve the current Azure workflow skip while the cold-cache representative path remains tracked by issue #65.
