+++ src/Arius.E2E.Tests/RepresentativeArchiveRestoreTests.cs
@@ -17,11 +17,11 @@ internal class RepresentativeArchiveRestoreTests
             return;
         }
 
-        //if (backend is AzureE2EBackendFixture)
-        //{
-        //    Skip.Unless(false, "Azure canonical representative workflow still includes the cold-cache restore path tracked by issue #65");
-        //    return;
-        //}
+        if (backend is AzureE2EBackendFixture)
+        {
+            Skip.Unless(false, "Azure canonical representative workflow still includes the cold-cache restore path tracked by issue #65");
+            return;
+        }
 
         var result = await RepresentativeWorkflowRunner.RunAsync(
             backend,
