+++ src/Arius.E2E.Tests/Workflows/Steps/ArchiveTierLifecycleStep.cs
@@ -22,23 +22,24 @@ internal sealed record ArchiveTierLifecycleStep(string Name, string TargetPath =
 
         await state.Fixture.MaterializeSourceAsync(state.Definition, sourceVersion, state.Seed);
 
-        var archiveResult = await RepresentativeWorkflowRunner.ArchiveAsync(
+        var archiveResult = await ArchiveStepSupport.ArchiveAsync(
             state.Fixture,
-            RepresentativeWorkflowRunner.CreateArchiveTierOptions(state.Fixture),
-            cancellationToken);
+            uploadTier: BlobTier.Archive,
+            cancellationToken: cancellationToken);
 
         archiveResult.Success.ShouldBeTrue($"{Name}: archive failed: {archiveResult.ErrorMessage}");
 
-        var tarChunkHash = await RepresentativeWorkflowRunner.PollForArchiveTierTarChunkAsync(azureBlobContainer, cancellationToken);
+        var tarChunkHash = await ArchiveTierStepSupport.PollForArchiveTierTarChunkAsync(azureBlobContainer, cancellationToken);
         tarChunkHash.ShouldNotBeNullOrWhiteSpace($"{Name}: expected at least one archive-tier tar chunk.");
 
-        var contentHashToBytes = await RepresentativeWorkflowRunner.ReadArchiveTierContentBytesAsync(
+        var contentHashToBytes = await ArchiveTierStepSupport.ReadContentBytesAsync(
             state.Fixture.LocalRoot,
             TargetPath);
 
         var firstEstimateCaptured = false;
         var firstTrackingBlobService = new CopyTrackingBlobService(azureBlobContainer);
-        var initialResult = await RepresentativeWorkflowRunner.CreateArchiveTierRestoreHandler(state.Fixture, state.Context, firstTrackingBlobService)
+        var initialRestoreHandler = await ArchiveTierStepSupport.CreateRestoreHandlerAsync(state.Fixture, state.Context, firstTrackingBlobService);
+        var initialResult = await initialRestoreHandler
             .Handle(new RestoreCommand(new RestoreOptions
             {
                 RootDirectory = state.Fixture.RestoreRoot,
@@ -57,14 +58,15 @@ internal sealed record ArchiveTierLifecycleStep(string Name, string TargetPath =
         initialResult.ChunksPendingRehydration.ShouldBeGreaterThan(0, $"{Name}: pending restore should report pending chunks.");
         initialResult.FilesRestored.ShouldBe(0, $"{Name}: pending restore should not restore files before rehydration is ready.");
 
-        var pendingRehydratedBlobCount = await RepresentativeWorkflowRunner.CountBlobsAsync(
+        var pendingRehydratedBlobCount = await ArchiveTierStepSupport.CountBlobsAsync(
             azureBlobContainer,
             BlobPaths.ChunksRehydrated,
             cancellationToken);
         pendingRehydratedBlobCount.ShouldBeGreaterThan(0, $"{Name}: pending restore should stage rehydrated chunk blobs.");
 
         var rerunTrackingBlobService = new CopyTrackingBlobService(azureBlobContainer);
-        var rerunResult = await RepresentativeWorkflowRunner.CreateArchiveTierRestoreHandler(state.Fixture, state.Context, rerunTrackingBlobService)
+        var rerunRestoreHandler = await ArchiveTierStepSupport.CreateRestoreHandlerAsync(state.Fixture, state.Context, rerunTrackingBlobService);
+        var rerunResult = await rerunRestoreHandler
             .Handle(new RestoreCommand(new RestoreOptions
             {
                 RootDirectory = state.Fixture.RestoreRoot,
@@ -76,12 +78,12 @@ internal sealed record ArchiveTierLifecycleStep(string Name, string TargetPath =
         rerunResult.Success.ShouldBeTrue($"{Name}: pending rerun failed: {rerunResult.ErrorMessage}");
         rerunTrackingBlobService.CopyCalls.Count.ShouldBe(0, $"{Name}: rerun should not issue duplicate rehydration copy requests.");
 
-        await RepresentativeWorkflowRunner.DeleteBlobsAsync(
+        await ArchiveTierStepSupport.DeleteBlobsAsync(
             azureBlobContainer,
             BlobPaths.ChunksRehydrated,
             cancellationToken);
 
-        await RepresentativeWorkflowRunner.SideloadRehydratedTarChunkAsync(
+        await ArchiveTierStepSupport.SideloadRehydratedTarChunkAsync(
             azureBlobContainer,
             tarChunkHash!,
             contentHashToBytes,
@@ -108,7 +110,7 @@ internal sealed record ArchiveTierLifecycleStep(string Name, string TargetPath =
             readyResult.Success.ShouldBeTrue($"{Name}: ready restore failed: {readyResult.ErrorMessage}");
             readyResult.ChunksPendingRehydration.ShouldBe(0, $"{Name}: ready restore should not leave pending rehydration chunks.");
 
-            await RepresentativeWorkflowRunner.AssertArchiveTierRestoreOutcomeAsync(
+            await ArchiveTierStepSupport.AssertRestoreOutcomeAsync(
                 state.Definition,
                 sourceVersion,
                 state.Seed,
