+++ src/Arius.E2E.Tests/Workflows/Steps/RestoreStep.cs
@@ -23,25 +23,13 @@ internal sealed record RestoreStep(
 
         Directory.CreateDirectory(state.Fixture.RestoreRoot);
 
-        var version = Target switch
-        {
-            WorkflowRestoreTarget.Previous => state.PreviousSnapshotVersion ?? throw new InvalidOperationException("Previous snapshot version is not available."),
-            _ => null,
-        };
+        var version = RestoreStepSupport.ResolveVersion(state, Target);
 
-        var result = await RepresentativeWorkflowRunner.RestoreAsync(
-            state.Fixture,
-            new RestoreOptions
-            {
-                RootDirectory = state.Fixture.RestoreRoot,
-                Overwrite = Overwrite,
-                Version = version,
-            },
-            cancellationToken);
+        var result = await RestoreStepSupport.RestoreAsync(state.Fixture, Overwrite, version, cancellationToken);
 
         result.Success.ShouldBeTrue($"{Name}: {result.ErrorMessage}");
 
-        await RepresentativeWorkflowRunner.AssertRestoreOutcomeAsync(
+        await RestoreStepSupport.AssertRestoreOutcomeAsync(
             state.Fixture,
             state.Definition,
             ExpectedVersion,
